WIPPY WIP - overhaul everything
