+++ automation/default_manager_params.json
@@ -2,6 +2,10 @@
     "data_directory": "~/openwpm/",
     "log_directory": "~/openwpm/",
     "output_format": "local",
+    "output_structured_format": "sqlite",
+    "output_unstructured_format": "leveldb",
+    "output_destination_type": "local",
+    "output_destination": "~/openwpm/",
     "database_name": "crawl-data.sqlite",
     "log_file": "openwpm.log",
     "failure_limit": null,
